% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fjson.R
\name{fjson}
\alias{fjson}
\title{Fast JSON reader}
\usage{
fjson(path = NULL, text = NULL, nThread = NULL, verbose = TRUE)
}
\arguments{
\item{path}{path to a JSON file.}

\item{text}{input data itself as a character vector of one or more lines, for example as returned by readLines().}

\item{nThread}{number of threads to use. By default it is set to the number of cores minus one.}

\item{verbose}{logical. Should progress indicators be supressed?}
}
\value{
data.table
}
\description{
fjson reads a JSON file into a data table.
}
\details{
fjson is a fast reader for JSON files. It is intnded for reading in JSON data that do not contain arrays, so it can be nicely organized into a table. The JSON file should have line breaks to separate each row of data.

It offers several advantages over existing packages:
\enumerate{
\item It is substantially faster, especially for large files. This is because the computationally intensive operations are parallelized using the parallel package, and are implemented in C++ using the Rcpp package.
\item Data type conversions are automatically done based on the parse_guess() function from the readr package.
\item It includes progress indicators.
}
}
\examples{
library(fjson)
df <- fjson(path = "file.json")
df <- fjson(text = c('{"time":"2019-01-01 00:00:00 UTC","character":"abc","integer_quoted":"30","integer_unquoted":30,"double_unquoted":30.333333}',
'{"character":"abcd","time":"2019-01-02 00:00:00 UTC","integer_quoted":"10","integer_unquoted":10,"double_unquoted":10.25}'))

}
\author{
Ariel Fridman
}
