% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fjson_parse_string}
\alias{fjson_parse_string}
\title{Reads a single row of JSON and returns a data frame with one row}
\usage{
fjson_parse_string(string)
}
\arguments{
\item{string}{A single string.}
}
\description{
Reads a single row of JSON and returns a data frame with one row
}
