% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fjson_parse_vector}
\alias{fjson_parse_vector}
\title{Reads multiple rows of JSON and returns a list of data frames}
\usage{
fjson_parse_vector(strings)
}
\arguments{
\item{strings}{A vector of strings.}
}
\description{
Reads multiple rows of JSON and returns a list of data frames
}
